==> analysis/04-group-analysis.Rmd <==
---
title: "Group Analysis"
author: ["Shixiang Wang", "Huimin Li", "Xuan Wang", "Minfang Song", "Xue-Song Liu (Corresponding author)"]
date: "`r Sys.Date()`"
output:
  rmdformats::readthedown:
    highlight: kate
    self_contained: false
    mathjax: true
    lightbox: true
    gallery: true
    toc: 3
bibliography: ref.bib
link-citations: yes
---

```{r knitr_init_04, echo=FALSE}
library(knitr)
library(rmdformats)
## Global options
options(max.print = "75")
opts_chunk$set(
  echo = TRUE,
  comment = "#>",
  collapse = TRUE,
  cache = FALSE,
  dpi = 300,
  warning = FALSE,
  message = FALSE,
  tidy = "styler"
)
opts_knit$set(width = 75)
```

This part has been divided into four parts to show the differences across sample groups defined by signature exposure.
It provides insight into the heterogeneity of prostate cancers and relative importance of two types of genomic variation signatures.


Load packages.

```{r, message=FALSE}
library(tidyverse)
library(sigminer)
```

Load the data.

```{r}
df.seqz = readRDS(file = "../output/df.seqz.RDS")
```

## Group comparison

```{r}
cols_to_features <- c(
  "IsMetastatic", "Stage", "GleasonScore",
  sort(c("Age", 
  "n_SBS",
  "n_INDEL", 
  "Ti_fraction", "Tv_fraction",
  "cnaBurden", "n_CNV", "n_Amp", "n_Del", "ploidy",
  "TDP score",
  "Chromoth_state",
  "MATH", "purity"))
)

feature_type <- c(rep("ca", 3L), rep("co", length(cols_to_features) - 3L))
```

```{r}
df.seqz$GleasonScore = as.character(df.seqz$GleasonScore)
df.seqz$GleasonScore = factor(df.seqz$GleasonScore, levels = as.character(6:10))

df.seqz2 = df.seqz
df.seqz2$cnv_enrich_sig = ifelse(
  !is.na(df.seqz2$cnv_enrich_sig),
  paste(df.seqz2$cnv_enrich_sig, "enriched"),
  NA
)

grp.cnv <- get_group_comparison(
  df.seqz2,
  col_group = "cnv_enrich_sig", # or "cnv_group"
  cols_to_compare = cols_to_features,
  type = feature_type
)
```

```{r}
plot.cnv <- show_group_comparison(
  grp.cnv,
  xlab = NA,
  method = "anova",
  text_angle_x = 60,
  text_hjust_x = 1,
  legend_position_ca = "right",
  label.x.npc = "center", 
  label.y.npc = "top",
  label = "p.format",
  hjust = 0.5
)
```

```{r, fig.width=9, fig.height=6}
cowplot::plot_grid(plotlist = plot.cnv$ca, align = "hv", ncol = 2)
```

```{r, include=FALSE}
p = cowplot::plot_grid(plotlist = plot.cnv$ca, align = "hv", ncol = 2)
ggsave(filename = "../figures/Figure5B.pdf", plot = p,
       width = 9, height = 6)
```


```{r, fig.width=12, fig.height=10}
cowplot::plot_grid(plotlist = plot.cnv$co, align = "hv", ncol = 4)
```

```{r, include=FALSE}
p = cowplot::plot_grid(plotlist = plot.cnv$co, align = "hv", ncol = 4)
ggsave(filename = "../figures/Figure_group_boxplot_cn.pdf", plot = p,
       width = 12, height = 10)
```

```{r}
df.seqz2$snv_enrich_sig = ifelse(
  !is.na(df.seqz2$snv_enrich_sig),
  paste(df.seqz2$snv_enrich_sig, "enriched"),
  NA
)

grp.snv <- get_group_comparison(
  df.seqz2,
  col_group = "snv_enrich_sig", # or "snv_group"
  cols_to_compare = cols_to_features,
  type = feature_type
)
```

```{r}
plot.snv <- show_group_comparison(
  grp.snv,
  xlab = NA,
  method = "anova",
  text_angle_x = 60,
  text_hjust_x = 1,
  legend_position_ca = "right",
  label.x.npc = "center", 
  label.y.npc = "top",
  label = "p.format",
  hjust = 0.5
)
```

```{r, fig.width=9, fig.height=6}
cowplot::plot_grid(plotlist = plot.snv$ca, align = "hv", ncol = 2)
```

```{r, include=FALSE}
p = cowplot::plot_grid(plotlist = plot.snv$ca, align = "hv", ncol = 2)
ggsave(filename = "../figures/Figure5C.pdf", plot = p,
       width = 9, height = 6)
```


```{r, fig.width=12, fig.height=10}
cowplot::plot_grid(plotlist = plot.snv$co, align = "hv", ncol = 4)
```

```{r, include=FALSE}
p = cowplot::plot_grid(plotlist = plot.snv$co, align = "hv", ncol = 4)
ggsave(filename = "../figures/Figure_group_boxplot_sbs.pdf", plot = p,
       width = 12, height = 10)
```

What if there are 5 SBS signatures and samples are divided into 5 groups?


```{r, message=FALSE}
load("../output/Sig5.Maf.RData")
sbs_grp = get_groups(Sig.SNV5, method = "consensus", match_consensus = TRUE)
sbs_grp$sample = ifelse(startsWith(sbs_grp$sample, "TCGA"),
                        substr(sbs_grp$sample, 1, 15),
                        sbs_grp$sample)
```

```{r}
df.seqz3 = dplyr::left_join(
  sbs_grp,
  df.seqz2,
  by = c("sample" = "tumor_Run")
)

df.seqz3$enrich_sig = ifelse(
  !is.na(df.seqz3$enrich_sig),
  paste(df.seqz3$enrich_sig, "enriched"),
  NA
)

grp.snv2 <- get_group_comparison(
  df.seqz3,
  col_group = "enrich_sig", # or "snv_group"
  cols_to_compare = cols_to_features,
  type = feature_type
)
```



```{r}
plot.snv2 <- show_group_comparison(
  grp.snv2,
  xlab = NA,
  method = "anova",
  text_angle_x = 60,
  text_hjust_x = 1,
  legend_position_ca = "right",
  label.x.npc = "center", 
  label.y.npc = "top",
  label = "p.format",
  hjust = 0.5
)
```

```{r, fig.width=9, fig.height=6}
cowplot::plot_grid(plotlist = plot.snv2$ca, align = "hv", ncol = 2)
```


```{r, fig.width=12, fig.height=10}
cowplot::plot_grid(plotlist = plot.snv2$co, align = "hv", ncol = 4)
```

## Group mapping

```{r}
df.map = df.seqz[, c("sample_type",
                     "cnv_enrich_sig",
                     "snv_enrich_sig")] %>% 
  na.omit()
names(df.map) = c("Sample type", "Enriched CN signature", "Enriched SBS signature")
```

```{r}
p = show_group_mapping(df.map, 
                   col_to_flow = "Sample type", 
                   cols_to_map = setdiff(colnames(df.map), "Sample type"),
                   fill_na = "NA", include_sig = TRUE)
p
```

```{r, include=FALSE}
ggsave(filename = "../figures/Figure5A.pdf", plot = p,
       width = 9, height = 7)
```


## Grouped signature exposure profile

We have showed the exposure profiles, here we add group annotation to them.

```{r}
load("../output/Sig.PRAD_TCGA_plus_dbGap_Maf.RData")
load("../output/Sig.CNV.seqz.W.RData")
```

### Copy number signatures

```{r}
map_df = df.seqz %>% 
  dplyr::select(cnv_enrich_sig, CNV_ID) %>% 
  dplyr::filter(!is.na(CNV_ID))

groups = map_df$cnv_enrich_sig
names(groups) = map_df$CNV_ID

p = show_sig_exposure(Sig.CNV.seqz.W,
                  groups = groups, 
                  #grp_order = , 
                  rm_space = TRUE, 
                  style = "cosmic",
                  rm_grid_line = FALSE,
                  rm_panel_border = TRUE,
                  grp_size = 7)
p
```

```{r, include=FALSE}
ggsave(filename = "../figures/Figure_grp_exposure_cn.pdf", plot = p,
       width = 12, height = 7)
```

### SBS signatures

```{r}
map_df = df.seqz %>% 
  dplyr::select(tumor_Run, snv_enrich_sig) %>% 
  dplyr::filter(!is.na(tumor_Run))

groups = map_df$snv_enrich_sig
names(groups) = map_df$tumor_Run

Sig.SNV2 = Sig.SNV

TCGA_INDEX = startsWith(colnames(Sig.SNV2$Exposure), "TCGA")
colnames(Sig.SNV2$Exposure)[TCGA_INDEX] = substr(colnames(Sig.SNV2$Exposure)[TCGA_INDEX], 1, 15)

p = show_sig_exposure(Sig.SNV2,
                  groups = groups,
                  cutoff = 2000,
                  rm_space = TRUE, 
                  style = "cosmic",
                  rm_grid_line = FALSE,
                  rm_panel_border = TRUE)
p
```

```{r, include=FALSE}
ggsave(filename = "../figures/Figure_grp_exposure_sbs.pdf", plot = p,
       width = 12, height = 7)
```

## Copy number profile of different groups

There are many differences across copy number groups. Here we go further showing some
classic cases by copy number profile.

```{r}
load("../output/CNV.seqz.RData")
load("../output/Sig.CNV.seqz.W.RData")
load("../output/CNV.seqz.tally.W.RData")
```

There are 5 copy number groups dominated by 4 copy number signatures.

```{r}
table(df.seqz$cnv_group, df.seqz$cnv_enrich_sig)
```

Get the groups and relative exposures and merge them.

```{r}
cn_group <- get_groups(Sig.CNV.seqz.W, method = "consensus", match_consensus = TRUE)
cn_group$enrich_sig[cn_group$group == "1"] = "Sig1"
cn_expo <- get_sig_exposure(Sig.CNV.seqz.W, type = "relative")

df <- dplyr::left_join(cn_group, cn_expo)
```

Now we can use data to find the cases with a specific signature enriched.

### Sig1 cases

The group with enriched Sig1 is dominated by Sig3, we need to take case of it.

```{r, fig.height=6, fig.width=14}
samps_to_show = df %>%
    dplyr::filter(enrich_sig == "Sig1") %>%
    dplyr::arrange(dplyr::desc(Sig1)) %>%
    dplyr::slice(1:6) %>% dplyr::pull(sample)

samps_to_show
                                                                                                                                                                                                                               
show_cn_profile(
  data = CNV.seqz, nrow = 3, ncol = 2, show_title = T,
  samples = samps_to_show 
)
```

Of note, only the segments with **high copy number** are contributed by Sig1.

### Sig2 cases

```{r, fig.height=6, fig.width=14}
samps_to_show = df %>%
    dplyr::filter(enrich_sig == "Sig2") %>%
    dplyr::arrange(dplyr::desc(Sig2)) %>%
    dplyr::slice(1:6) %>% dplyr::pull(sample)

samps_to_show

show_cn_profile(
  data = CNV.seqz, nrow = 3, ncol = 2, show_title = T,
  samples = samps_to_show
)
```

### Sig3 cases

```{r,fig.height=6, fig.width=14}
samps_to_show = df %>%
    dplyr::filter(enrich_sig == "Sig3") %>%
    dplyr::arrange(dplyr::desc(Sig3)) %>%
    dplyr::slice(1:6) %>% dplyr::pull(sample)

samps_to_show

show_cn_profile(
  data = CNV.seqz, nrow = 3, ncol = 2, show_title = T,
  samples = samps_to_show
)
```

### Sig4 cases

```{r fig.height=6, fig.width=14}
samps_to_show = df %>%
    dplyr::filter(enrich_sig == "Sig4") %>%
    dplyr::arrange(dplyr::desc(Sig4)) %>%
    dplyr::slice(1:6) %>% dplyr::pull(sample)

samps_to_show

show_cn_profile(
  data = CNV.seqz, nrow = 3, ncol = 2, show_title = T,
  samples = samps_to_show
)
```

### Sig5 cases

```{r fig.height=6, fig.width=14}
samps_to_show = df %>%
    dplyr::filter(enrich_sig == "Sig5") %>%
    dplyr::arrange(dplyr::desc(Sig5)) %>%
    dplyr::slice(1:6) %>% dplyr::pull(sample)

samps_to_show

show_cn_profile(
  data = CNV.seqz, nrow = 3, ncol = 2, show_title = T,
  samples = samps_to_show
)
```

Sig5 enriched case with alterations.

```{r fig.height=6, fig.width=14}
samps_to_show = df %>%
    dplyr::filter(enrich_sig == "Sig5") %>%
    dplyr::left_join(as.data.frame(CNV.seqz.tally.W$nmf_matrix) %>%
                tibble::rownames_to_column("sample") %>%
                dplyr::select(sample, `BPArm[2]`, `BPArm[1]`)) %>%
    dplyr::mutate(BK = `BPArm[2]` + `BPArm[1]`) %>%
    dplyr::arrange(desc(BK)) %>%
    dplyr::slice(1:6) %>% dplyr::pull(sample)

samps_to_show

show_cn_profile(
  data = CNV.seqz, nrow = 3, ncol = 2, show_title = T,
  samples = samps_to_show
)
```

### Selected cases

Select one sample for one case above.

```{r fig.height=10, fig.width=7}
p = show_cn_profile(
  data = CNV.seqz, nrow = 5, ncol = 1, show_title = T,
  samples = c("5115056-SRR8311885", "5115615-SRR8311749", "TCGA-KK-A7B4-01", "TCGA-KK-A59X-01", "TCGA-KC-A7FE-01")
)
p
```


```{r, include=FALSE}
ggsave(filename = "../figures/Figure_cn_profile_for_selected_cases.pdf", plot = p,
       width = 8, height = 10)
```


```{r, eval=FALSE}
pdf("Selected_cases_circos_heatmap.pdf", width = 2.5, height = 13)
samps = c("5115056-SRR8311885", "5115615-SRR8311749", "TCGA-KK-A7B4-01", "TCGA-KK-A59X-01", "TCGA-KC-A7FE-01")

opar = par(no.readonly = TRUE)
print(par("mar"))
par(mar = rep(0, 4))

layout(matrix(1:5, ncol = 1))
for (samp in samps) {
  circlize::circos.par(track.margin=c(0,0)) 
  show_cn_circos(CNV.seqz, samples = samp, show_title = FALSE, 
               col = circlize::colorRamp2(c(1, 2, 4), c("blue", "white", "red")),
               line_lwd = par("lwd") / 4)
}
layout(1)
par(opar)
dev.off()
```


```{r fig.height=12, fig.width=8}
# p_gg = show_cn_profile(
#   data = CNV.seqz, nrow = 6, ncol = 1, show_title = T,
#   samples = samps
# )
# 
# p_base = function() {
#   opar = par(no.readonly = TRUE)
#   par(mar = c(0, 0, 0, 0))
#   on.exit(par(opar))
#   
#   layout(matrix(1:5, ncol = 1))
#   show_cn_circos(CNV.seqz, samples = samps, show_title = FALSE, line_lwd = par("lwd") / 4)
#   layout(1)
# }

# cowplot::plot_grid(
#   p_gg,
#   p_base,
#   rel_widths = c(3, 1),
#   scale = c(1.0, 0.9)
# )
```
