==> analysis/05-survival-analysis.Rmd <==
---
title: "Survival Analysis"
author: ["Shixiang Wang", "Huimin Li", "Xuan Wang", "Minfang Song", "Xue-Song Liu (Corresponding author)"]
date: "`r Sys.Date()`"
output:
  rmdformats::readthedown:
    highlight: kate
    self_contained: false
    mathjax: true
    lightbox: true
    gallery: true
    toc: 3
bibliography: ref.bib
link-citations: yes
---

```{r knitr_init_05, echo=FALSE}
library(knitr)
library(rmdformats)
## Global options
options(max.print = "75")
opts_chunk$set(
  echo = TRUE,
  comment = "#>",
  collapse = TRUE,
  cache = TRUE,
  dpi = 300,
  warning = FALSE,
  message = FALSE,
  tidy = "styler"
)
opts_knit$set(width = 75)
```

To check the clinical relavance of signatures, this part we do survival analysis with Cox model.

## Prepare

Load package **ezcox** for batch Cox analysis.

```{r}
library(ezcox)
library(ggplot2)
```

Load the data.

```{r}
df.seqz = readRDS(file = "../output/df.seqz.RDS")
surv_df <- readRDS(file = "../data/PRAD_Survival.rds")
```

Clean the data to generate a data table for survival analysis.


```{r}
range01 <- function(x, ...) {
  (x - min(x, ...)) / (max(x, ...) - min(x, ...))
}

cols_to_sigs.seqz <- c(paste0("CN-Sig", 1:5), paste0("SBS-Sig", 1:3))


surv_dt <- dplyr::inner_join(df.seqz %>%
                               dplyr::select(-c("Study", "subject_id", "tumor_body_site",
                                                "tumor_Run", "normal_Run", "CNV_ID",
                                                "Fusion")),
                             surv_df, by = c("PatientID" = "sample"))

dup_ids = which(duplicated(surv_dt$PatientID))
# Remove duplicated records
surv_dt = surv_dt[-dup_ids, ]

# Scale the signature exposure by dividing 10
# Scale some variables to 0-20.
surv_dt = surv_dt %>%
  dplyr::select(-Tv_fraction) %>% 
  dplyr::mutate(cnaBurden = 20 * cnaBurden,
                Stage = as.character(Stage) %>% factor(levels = c("T2", "T3", "T4"))) %>%
  dplyr::mutate_at(cols_to_sigs.seqz, ~./10) %>% 
  dplyr::mutate_at(c("Ti_fraction", "purity"), ~ 20 * range01(., na.rm = TRUE)) %>% 
  dplyr::mutate(OS.time = OS.time / 30.5,
                PFI.time = PFI.time / 30.5)


# Scale the signature exposure to 0-20.
# To better understand the effect of signature exposure on clinical events,
# here we scale them into range 0-20. 1 increase means 5% increase of signature exposure.
#
# Scale the CNA burden to 0-20.
#
# keep only ti_fraction, tv result will be 1/ti_result
# also scale it to 0-20
# also scale purity to 0-20
# surv_dt = surv_dt %>%
#   dplyr::select(-Tv_fraction) %>% 
#   dplyr::mutate(cnaBurden = 20 * cnaBurden,
#                 Stage = as.character(Stage) %>% factor(levels = c("T2", "T3", "T4"))) %>%
#   dplyr::mutate_at(c(cols_to_sigs.seqz, "Ti_fraction", "purity"), ~ 20 * range01(., na.rm = TRUE))

saveRDS(surv_dt, file = "../output/PRAD_merged_survival_dt_seqz.rds")
```

## OS analysis fo signatures

```{r}
############
surv_dt <- readRDS(file = "../output/PRAD_merged_survival_dt_seqz.rds")
# surv_dt <- surv_dt %>% dplyr::filter(startsWith(PatientID, "TCGA"))

cols_to_sigs.seqz <- c(paste0("CN-Sig", 1:5), paste0("SBS-Sig", 1:3))

# Unvariable analysis
p = show_forest(surv_dt, 
                covariates = cols_to_sigs.seqz,
            time = "OS.time", status = "OS", merge_models = TRUE)
p
```

```{r, include=FALSE}
ggsave(filename = "../figures/Figure_OS_sig.pdf", plot = p,
       width = 7, height = 5)
```


```{r, fig.height=10, fig.width=7}
# Multivariable analysis
# p = show_forest(surv_dt, covariates = cols_to_sigs.seqz, controls = c("GleasonScore", "purity", "Stage"),
#             time = "OS.time", status = "OS", merge_models = TRUE, drop_controls = TRUE)
p = show_forest(surv_dt,
                covariates = cols_to_sigs.seqz, controls = c("GleasonScore", "purity", "Stage"),
            time = "OS.time", status = "OS")
p
```

```{r, include=FALSE}
ggsave(filename = "../figures/Figure_OS_sig_multivar.pdf", plot = p,
       width = 7, height = 10)
```

## PFS analysis for signatures


```{r}
p = show_forest(surv_dt, 
                covariates = cols_to_sigs.seqz,
            time = "PFI.time", status = "PFI", merge_models = TRUE)
p
```

```{r, include=FALSE}
ggsave(filename = "../figures/Figure_PFS_sig.pdf", plot = p,
       width = 7, height = 5)
```

```{r, fig.height=10, fig.width=7}
# Multivariable analysis
p = show_forest(surv_dt,
                covariates = cols_to_sigs.seqz, controls = c("GleasonScore", "purity", "Stage"),
            time = "PFI.time", status = "PFI")
p
```

```{r, include=FALSE}
ggsave(filename = "../figures/Figure_PFS_sig_multivar.pdf", plot = p,
       width = 7, height = 10)
```

## OS analysis for features

```{r}
cols_to_features <- c(
  "Age", "Stage", "GleasonScore",
  "n_SBS", 
  "n_INDEL",
  "n_CNV", "n_Amp", "n_Del", 
  "cnaBurden",
  "Ti_fraction",
  "TDP score",
  "Chromoth_state",
  "MATH",
  "purity",
  "ploidy"
)

p = show_forest(surv_dt,
            covariates = cols_to_features, controls = NULL,
            time = "OS.time", status = "OS", merge_models = TRUE, limits = log(c(0.5, 10)))
p
```

```{r, include=FALSE}
ggsave(filename = "../figures/Figure_OS_features.pdf", plot = p,
       width = 7, height = 6)
```

## PFS analysis for features


```{r}
p = show_forest(surv_dt,
            covariates = cols_to_features, controls = NULL,
            time = "PFI.time", status = "PFI", merge_models = TRUE, limits = c(log(0.5), log(5)))
p
```

```{r, include=FALSE}
ggsave(filename = "../figures/Figure_PFS_features.pdf", plot = p,
       width = 7, height = 6)
```


## K-M curves

```{r}
library(survival)
library(survminer)

legend_2 <- ggplot2::guides(color = ggplot2::guide_legend(nrow = 2, byrow = TRUE))
```

```{r}
fit <- survfit(Surv(OS.time, OS) ~ 1, data = subset(surv_dt, sample_type == "Primary"))
ggsurvplot(fit, palette = "aaas", xlab = "Time (months)")$plot
```

```{r}
fit <- survfit(Surv(OS.time, OS) ~ 1, data = subset(surv_dt, sample_type != "Primary"))
ggsurvplot(fit, palette = "aaas", xlab = "Time (months)")$plot
```

```{r}
fit <- survfit(Surv(PFI.time, PFI) ~ 1, data = subset(surv_dt, sample_type == "Primary"))
ggsurvplot(fit, palette = "aaas", xlab = "Time (months)", ylab = "Disease progression probability")$plot
```

```{r}
fit <- survfit(Surv(OS.time, OS) ~ cnv_enrich_sig, data = subset(surv_dt, sample_type == "Primary"))
ggsurvplot(fit, palette = "aaas", xlab = "Time (months)", pval = TRUE)$plot + legend_2
```

```{r}
fit <- survfit(Surv(OS.time, OS) ~ cnv_enrich_sig, data = subset(surv_dt, sample_type != "Primary"))
ggsurvplot(fit, palette = "aaas", xlab = "Time (months)", pval = TRUE)$plot + legend_2
```

```{r}
fit <- survfit(Surv(OS.time, OS) ~ snv_enrich_sig, data = subset(surv_dt, sample_type == "Primary"))
ggsurvplot(fit, palette = "aaas", xlab = "Time (months)", pval = TRUE)$plot + legend_2
```

```{r}
fit <- survfit(Surv(OS.time, OS) ~ snv_enrich_sig, data = subset(surv_dt, sample_type != "Primary"))
ggsurvplot(fit, palette = "aaas", xlab = "Time (months)", pval = TRUE)$plot + legend_2
```

```{r}
fit <- survfit(Surv(PFI.time, PFI) ~ cnv_enrich_sig, data = subset(surv_dt, sample_type == "Primary"))
ggsurvplot(fit, palette = "aaas", xlab = "Time (months)", ylab = "Disease progression probability",
           pval = TRUE)$plot + legend_2
```

```{r}
fit <- survfit(Surv(PFI.time, PFI) ~ snv_enrich_sig, data = subset(surv_dt, sample_type == "Primary"))
ggsurvplot(fit, palette = "aaas", xlab = "Time (months)", ylab = "Disease progression probability",
           pval = TRUE)$plot + legend_2
```

No PFI data for metastatic tumors.
